consolidate middlewares
