+++ services/horizon/internal/middleware.go
@@ -9,9 +9,47 @@ import (
 	"github.com/go-chi/chi"
 	"github.com/go-chi/chi/middleware"
 	chimiddleware "github.com/go-chi/chi/middleware"
+	"github.com/stellar/go/services/horizon/internal/errors"
+	"github.com/stellar/go/services/horizon/internal/httpx"
 	"github.com/stellar/go/services/horizon/internal/render"
+	"github.com/stellar/go/support/context/requestid"
 	"github.com/stellar/go/support/log"
+	"github.com/stellar/go/support/render/problem"
 )
+
+// Adds the "app" context into every request, so that subsequence middleware
+// or handlers can retrieve a horizon.App instance
+func (app *App) Middleware(h http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		ctx := app.Context(r.Context())
+		h.ServeHTTP(w, r.WithContext(ctx))
+	})
+}
+
+// requestCacheHeadersMiddleware adds caching headers to each response.
+func requestCacheHeadersMiddleware(h http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		// Before changing this read Stack Overflow answer about staled request
+		// in older versions of Chrome:
+		// https://stackoverflow.com/questions/27513994/chrome-stalls-when-making-multiple-requests-to-same-resource
+		w.Header().Set("Cache-Control", "no-cache, no-store, max-age=0")
+		h.ServeHTTP(w, r)
+	})
+}
+
+func contextMiddleware(parent context.Context) func(next http.Handler) http.Handler {
+	return func(next http.Handler) http.Handler {
+		fn := func(w http.ResponseWriter, r *http.Request) {
+			ctx := r.Context()
+			ctx = requestid.ContextFromChi(ctx)
+			ctx, cancel := httpx.RequestContext(ctx, w, r)
+
+			defer cancel()
+			next.ServeHTTP(w, r.WithContext(ctx))
+		}
+		return http.HandlerFunc(fn)
+	}
+}
 
 const (
 	clientNameHeader    = "X-Client-Name"
@@ -52,7 +90,6 @@ func getClientData(r *http.Request, headerName string) string {
 	}
 
 	value = r.URL.Query().Get(headerName)
-
 	if value == "" {
 		value = "undefined"
 	}
@@ -97,4 +134,55 @@ func logEndOfRequest(ctx context.Context, r *http.Request, duration time.Duratio
 		"status":         mw.Status(),
 		"streaming":      streaming,
 	}).Info("Finished request")
+}
+
+func (web *Web) RateLimitMiddleware(next http.Handler) http.Handler {
+	if web.rateLimiter == nil {
+		return next
+	}
+	return web.rateLimiter.RateLimit(next)
+}
+
+// RecoverMiddleware helps the server recover from panics.  It ensures that
+// no request can fully bring down the horizon server, and it also logs the
+// panics to the logging subsystem.
+func RecoverMiddleware(h http.Handler) http.Handler {
+	fn := func(w http.ResponseWriter, r *http.Request) {
+		ctx := r.Context()
+		defer func() {
+
+			if rec := recover(); rec != nil {
+				err := errors.FromPanic(rec)
+				errors.ReportToSentry(err, r)
+				problem.Render(ctx, w, err)
+			}
+		}()
+
+		h.ServeHTTP(w, r)
+	}
+
+	return http.HandlerFunc(fn)
+}
+
+// Middleware that records metrics.
+//
+// It records success and failures using a meter, and times every request
+func requestMetricsMiddleware(h http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		app := AppFromContext(r.Context())
+		mw := middleware.NewWrapResponseWriter(w, r.ProtoMajor)
+
+		app.web.requestTimer.Time(func() {
+			h.ServeHTTP(mw.(http.ResponseWriter), r)
+		})
+
+		if 200 <= mw.Status() && mw.Status() < 400 {
+			// a success is in [200, 400)
+			app.web.successMeter.Mark(1)
+		} else if 400 <= mw.Status() && mw.Status() < 600 {
+			// a success is in [400, 600)
+			app.web.failureMeter.Mark(1)
+		}
+
+	})
 }
