--- services/horizon/internal/middleware_rate_limit.go
@@ -1,12 +1,0 @@
-package horizon
-
-import (
-	"net/http"
-)
-
-func (web *Web) RateLimitMiddleware(next http.Handler) http.Handler {
-	if web.rateLimiter == nil {
-		return next
-	}
-	return web.rateLimiter.RateLimit(next)
-}
