--- services/horizon/internal/middleware_recover.go
@@ -1,29 +1,0 @@
-package horizon
-
-import (
-	"net/http"
-
-	"github.com/stellar/go/services/horizon/internal/errors"
-	"github.com/stellar/go/support/render/problem"
-)
-
-// RecoverMiddleware helps the server recover from panics.  It ensures that
-// no request can fully bring down the horizon server, and it also logs the
-// panics to the logging subsystem.
-func RecoverMiddleware(h http.Handler) http.Handler {
-	fn := func(w http.ResponseWriter, r *http.Request) {
-		ctx := r.Context()
-		defer func() {
-
-			if rec := recover(); rec != nil {
-				err := errors.FromPanic(rec)
-				errors.ReportToSentry(err, r)
-				problem.Render(ctx, w, err)
-			}
-		}()
-
-		h.ServeHTTP(w, r)
-	}
-
-	return http.HandlerFunc(fn)
-}
